Extensions/RandomValue/PythonRandomValue/requirements.txt
# debugpy

--find-links file:///C:/ProgramData/Beckhoff/PythonPackages
beckhoff-tchmi-extensionapi  # must be installed via TcPkg by the Python.Beckhoff.tchmi.extensionapi package.
